+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add responsive menu component [#674](https://github.com/azavea/echo-locator/pull/674)
 - Add bar charts: Neighborhood Criteria and Commute Time [#675](https://github.com/azavea/echo-locator/pull/675)
 - Add Neighborhood Card [#678](https://github.com/azavea/echo-locator/pull/678)
+- Add input components: Text, Number, Checkbox, Slider [#677](https://github.com/azavea/echo-locator/pull/677)
 
 ### Changed